# Conflicts:
#	src/dbus_api.rs

+++ src/engine.rs
@@ -1,3 +1,7 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
 use types::StratisResult;
 
 pub trait Pool {
@@ -10,7 +14,7 @@ pub trait Engine {
     fn create_pool(&self,
                    name: &str,
                    blockdev_paths: &[&str],
-                   raid_level: i32)
+                   raid_level: u16)
                    -> StratisResult<Box<Pool>>;
 
     fn destroy_pool(&self, name: &str) -> StratisResult<()>;
