+++ src/sim_engine.rs
@@ -1,3 +1,6 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 use types::StratisResult;
 use engine::{Engine, Pool};
@@ -17,7 +20,7 @@ impl Engine for SimEngine {
     fn create_pool(&self,
                    name: &str,
                    blockdev_paths: &[&str],
-                   raid_level: i32)
+                   raid_level: u16)
                    -> StratisResult<Box<Pool>> {
         println!("sim: pool created");
